test: 핸들러테스트 패키지 변경
