+++ tomcat/src/test/java/com/techcourse/handler/NotFoundHandlerTest.java
@@ -1,8 +1,7 @@
-package org.apache.coyote.http11;
+package com.techcourse.handler;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.techcourse.handler.NotFoundHandler;
 import java.io.File;
 import java.net.URL;
 import java.nio.file.Files;
